Taskrc changed to be specific to osx/linux

+++ zshrc
@@ -4,34 +4,43 @@ ZSH=$HOME/.oh-my-zsh
 # Theme
 ZSH_THEME="dieter"
 
+# ZSH
+source $ZSH/oh-my-zsh.sh
+
 # Various
-alias news="newsbeuter"
 alias irc='irssi -c im.bitlbee.org'
 alias matrix="cmatrix -bs"
-alias wifi="wicd-curses"
-alias m="ncmpcpp"
 alias h="htop"
-alias bat="acpi -b"
 alias sl="ls"
 alias v="vim"
 alias df="df -h --total"
-alias du="du -h -d 2 -c"
 alias list="ls -alog1"
 alias ll="ls -alGh"
 alias la="ls -a"
 alias l="clear && pwd && ls -FGl"
 alias ..="cd .."
 alias u="cd .. && l"
+if [[ $HOME == "/home/thinkbot" ]]; then
+    alias du="du -h -d 2 -c -a"
+    alias news="newsbeuter"
+    alias bat="acpi -b"
+    alias wifi="wicd-curses"
+    alias m="ncmpcpp"
+elif [[ $HOME == "/Users/mrtwiddletoes" ]]; then
+    alias du="du -h -d 2 -c"
+fi
 
 # Suffix Aliases
-alias -s pdf=evince
 alias -s txt=vim
 alias -s doc=libreoffice
 alias -s odf=libreoffice
+if [[ $HOME == "/home/thinkbot" ]]; then
+    alias -s pdf=evince
+fi
 
 # Directories
 hash -d src=~/src
-hash -d dotfiles=~/dotfiles
+hash -d dot=~/dotfiles
 
 # Permissions
 alias privatize="sudo chmod 700"
@@ -40,19 +49,26 @@ alias publicize="sudo chmod 777"
 # Configuration
 alias zshrc="vim ~/dotfiles/zshrc"
 alias vimrc="vim ~/dotfiles/vimrc"
-alias font="setfont /usr/share/kbd/consolefonts/Lat2-Terminus16.psfu.gz"
+if [[ $HOME == "/home/thinkbot" ]]; then
+    alias font="setfont /usr/share/kbd/consolefonts/Lat2-Terminus16.psfu.gz"
+fi
 
 # Package Manager
-alias upd="sudo pacman -Syy"
-alias upg="sudo pacman -Syu"
 alias gupd="gem update"
-alias bupd="brew update"
-alias bupg="brew upgrade"
+if [[ $HOME == "/Users/mrtwiddletoes" ]]; then
+    alias upd="brew update"
+    alias upg="brew upgrade"
+elif [[ $HOME == "/home/thinkbot" ]]; then
+    alias upd="sudo pacman -Syy"
+    alias upg="sudo pacman -Syu"
+fi
 
 # Hardware
-alias sus="sudo pm-suspend"
-alias sd="sudo shutdown -h now"
-alias rs="sudo shutdown -r now"
+if [[ $HOME == "/home/thinkbot" ]]; then
+    alias sus="sudo pm-suspend"
+    alias sd="sudo shutdown -h now"
+    alias rs="sudo shutdown -r now"
+fi
 
 # Extended History
 EXTENDED_HISTORY="true"
@@ -67,31 +83,19 @@ elif [[ $HOME == "/home/thinkbot" ]]; then
     plugins=(git extract git-flow pip fasd ruby gem github node npm nyan python taskwarrior vi-mode archlinux)
 fi
 
-# ZSH
-source $ZSH/oh-my-zsh.sh
-
-# Task Warrior Tab Completion
-fpath=($fpath /usr/local/share/doc/task/scripts/zsh)
-autoload -Uz compinit
-compinit
-zstyle ':completion:*' verbose yes
-zstyle ':completion:*:descriptions' format '%U%B%d%b%u'
-zstyle ':completion:*' group-name ''
-autoload colors
-zstyle ':completion:*:*:task:*:arguments' list-colors "=(#b) #([^-]#)*=$color[cyan]=$color[bold];$color[blue]" 
-zstyle ':completion:*:*:task:*:default' list-colors "=(#b) #([^-]#)*=$color[cyan]=$color[green]" 
-zstyle ':completion:*:*:task:*:values' list-colors "=(#b) #([^-]#)*=$color[cyan]=$color[bold];$color[red]" 
-zstyle ':completion:*:*:task:*:commands' list-colors "=(#b) #([^-]#)*=$color[cyan]=$color[yellow]" 
-
 # Path
 export PATH=~/bin:/usr/lib/lightdm/lightdm:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:.:/usr/local/MATLAB/R2011a/bin:/home/thinkbot/.gem/ruby/1.9.1/bin:/root/.gem/ruby/1.9.1/bin
 
 # XDG
-XDG_CONFIG_HOME=~/.config
-XDG_CONFIG_DIRS=~/.config:$XDG_CONFIG_DIRS
+if [[ $HOME == "/home/thinkbot" ]]; then
+    XDG_CONFIG_HOME=~/.config
+    XDG_CONFIG_DIRS=~/.config:$XDG_CONFIG_DIRS
+fi
 
+# No corrections
 unsetopt correctall
 
+# Automatically start tmux if our term is xterm or rxvt
 case $TERM in
 xterm*|rxvt*)
     if which tmux 2>&1 >/dev/null; then
